src/backend/django_src/requirements.txt
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.0.3
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.1
pyzmq==25.1.2
requests==2.27.1
six==1.16.0
spotipy==2.19.0
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.8
wcwidth==0.2.13
Werkzeug==2.0.2
